Completed assignment

+++ Binomial.java
@@ -22,12 +22,12 @@
 		}
 
 		// Computes the Binomial function, efficiently
-		public static int binomial(int n, int k) {
+		public static long binomial(int n, int k) {
 			//// This function creates a 2D array, say memo, / and then initializes all its
 			/// elements to -1. / It then calls binomial(n, k, memo), which does all the
 			/// heavy lifiting.
 			// Replace the following statement with your code.
-			int[][] memo = new int[n + 1][k + 1];
+			long[][] memo = new long [n + 1][k + 1];
 			for (int i = 0; i <= n; i++) {
 				for (int j = 0; j <= k; j++) {
 					memo[i][j] = -1;
@@ -36,7 +36,7 @@
 			return binomial(n, k, memo);
 		}
 
-		private static int binomial(int n, int k, int[][] memo) {
+		private static long binomial(int n, int k, long[][] memo) {
 			if (memo[n][k] != -1) {
 				return memo[n][k];
 			}
